+++ biaslyze/concepts/concepts_en.py
@@ -1,63 +1,7 @@
-"""Dict of concepts with related keywords.
+"""Dict of concepts with related keywords for the English language.
 
 Function abbreviations can be found here: https://v2.spacy.io/api/annotation#pos-tagging 
 """
-CONCEPTS_DE = {
-    "gender": [
-        {"keyword": "sie", "function": ["subject"], "category": "PRON"},
-        {"keyword": "er", "function": ["subject"], "category": "PRON"},
-        {"keyword": "ihr", "function": ["object"], "category": "PRON"},
-        {"keyword": "ihn", "function": ["object"], "category": "PRON"},
-        {"keyword": "ihre", "function": ["object"], "category": "PRON"},
-        {"keyword": "ihm", "function": ["object"], "category": "PRON"},
-        {"keyword": "seine", "function": ["object"], "category": "PRON"},
-        {"keyword": "seinen", "function": ["object"], "category": "PRON"},
-        {"keyword": "seiner", "function": ["object"], "category": "PRON"},
-        {"keyword": "seines", "function": ["object"], "category": "PRON"},
-        {"keyword": "ihres", "function": ["object"], "category": "PRON"},
-        {"keyword": "ihren", "function": ["object"], "category": "PRON"},
-        {"keyword": "ihrem", "function": ["object"], "category": "PRON"},
-        {"keyword": "ihnen", "function": ["object"], "category": "PRON"},
-    ],
-    "religion": [
-        {"keyword": "christ", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "christin", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "christen", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "christentum", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "christlich", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "christliche", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "christlichen", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "christlicher", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "christliches", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "christinnen", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "christi", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "christus", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "jude", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "jüdin", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "jüdisch", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "jüdische", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "jüdischen", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "jüdischer", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "jüdisches", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "jüdisches", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "judentum", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "muslim", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "muslima", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "muslimisch", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "muslimische", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "muslimischen", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "muslimischer", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "muslimisches", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "muslime", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "islam", "function": ["NOUN"], "category": "religion"},
-        {"keyword": "islamisch", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "islamische", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "islamischen", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "islamischer", "function": ["ADJ"], "category": "religion"},
-        {"keyword": "islamisches", "function": ["ADJ"], "category": "religion"},
-    ],        
-}
-
 
 CONCEPTS_EN = {
     "nationality": [
